+++ packages/components/Link/src/Link.tsx
@@ -7,7 +7,7 @@ import { compose, IUseComposeStyling } from '@uifabricshared/foundation-compose'
 import { ILinkProps, ILinkSlotProps, ILinkState, ILinkRenderData, IWithLinkOptions, linkName, ILinkType } from './Link.types';
 import { settings } from './Link.settings';
 import { foregroundColorTokens, textTokens, borderTokens } from '@fluentui-react-native/tokens';
-import { useAsPressable, useKeyCallback, useOnPressWithFocus, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
+import { useAsPressable, useKeyUpProps, useOnPressWithFocus, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
 import { ISlots, withSlots } from '@uifabricshared/foundation-composable';
 import { IViewProps } from '@fluentui-react-native/adapters';
@@ -34,7 +34,7 @@ export function useAsLink(userProps: IWithLinkOptions<IViewProps>, ref: React.Re
   const linkOnPressWithFocus = useOnPressWithFocus(ref, linkOnPress);
 
   const pressable = useAsPressable({ onPress: linkOnPressWithFocus, ...rest });
-  const onKeyUp = useKeyCallback(linkOnPress, ' ', 'Enter');
+  const onKeyUpProps = useKeyUpProps(linkOnPress, ' ', 'Enter');
 
   const newState = {
     ...pressable.state,
@@ -44,7 +44,7 @@ export function useAsLink(userProps: IWithLinkOptions<IViewProps>, ref: React.Re
   const newProps = {
     ...userProps,
     ...pressable.props,
-    onKeyUp,
+    ...onKeyUpProps,
   };
 
   return [newProps, newState];