+++ packages/experimental/Button/src/ButtonTokens.ts
@@ -12,12 +12,19 @@ export const defaultButtonTokens: TokenSettings<ButtonTokens, Theme> = () =>
       padding: globalTokens.spacing.sNudge - globalTokens.stroke.width.thin,
       borderWidth: globalTokens.stroke.width.thin,
       iconSize: 16,
+      focused: {
+        borderWidth: 0,
+        padding: globalTokens.spacing.sNudge,
+      },
       hasContent: {
         minWidth: 96,
-        paddingHorizontal: globalTokens.spacing.l - globalTokens.stroke.width.thin,
+        paddingHorizontal: globalTokens.spacing.m - globalTokens.stroke.width.thin,
         variant: 'bodySemibold',
         hasIcon: {
           spacingIconContent: globalTokens.spacing.sNudge,
+        },
+        focused: {
+          paddingHorizontal: globalTokens.spacing.m,
         },
       },
     },
@@ -25,12 +32,19 @@ export const defaultButtonTokens: TokenSettings<ButtonTokens, Theme> = () =>
       padding: globalTokens.spacing.xs - globalTokens.stroke.width.thin,
       borderWidth: globalTokens.stroke.width.thin,
       iconSize: 16,
+      focused: {
+        borderWidth: 0,
+        padding: globalTokens.spacing.xs,
+      },
       hasContent: {
         minWidth: 64,
-        paddingHorizontal: globalTokens.spacing.m - globalTokens.stroke.width.thin,
+        paddingHorizontal: globalTokens.spacing.s - globalTokens.stroke.width.thin,
         variant: 'secondaryStandard',
         hasIcon: {
           spacingIconContent: globalTokens.spacing.xs,
+        },
+        focused: {
+          paddingHorizontal: globalTokens.spacing.s,
         },
       },
     },
@@ -38,12 +52,19 @@ export const defaultButtonTokens: TokenSettings<ButtonTokens, Theme> = () =>
       padding: globalTokens.spacing.s - globalTokens.stroke.width.thin,
       borderWidth: globalTokens.stroke.width.thin,
       iconSize: 20,
+      focused: {
+        borderWidth: 0,
+        padding: globalTokens.spacing.s,
+      },
       hasContent: {
         minWidth: 96,
         paddingHorizontal: globalTokens.spacing.l - globalTokens.stroke.width.thin,
         variant: 'subheaderSemibold',
         hasIcon: {
           spacingIconContent: globalTokens.spacing.sNudge,
+        },
+        focused: {
+          paddingHorizontal: globalTokens.spacing.l,
         },
       },
     },