+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,50 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Wed, 17 Nov 2021 19:28:07 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Dec 2021 22:06:58 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.10.3
+
+Fri, 17 Dec 2021 22:06:58 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.4
+
+## 0.10.2
+
+Fri, 17 Dec 2021 19:53:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.3
+
+## 0.10.1
+
+Fri, 17 Dec 2021 01:26:42 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.9.2
+- Bump @fluentui-react-native/framework to v0.7.2
+- Bump @fluentui-react-native/text to v0.11.2
+- Bump @fluentui-react-native/theming-utils to v0.8.2
+- Bump @uifabricshared/foundation-compose to v1.11.2
+
+## 0.10.0
+
+Thu, 16 Dec 2021 19:10:27 GMT
+
+### Minor changes
+
+- Adding E2E functional testing for rest of phase 2 components (safreibe@microsoft.com)
+- Bump @fluentui-react-native/default-theme to v0.9.1
+- Bump @fluentui-react-native/framework to v0.7.1
+- Bump @fluentui-react-native/text to v0.11.1
+- Bump @fluentui-react-native/theming-utils to v0.8.1
+- Bump @fluentui-react-native/tokens to v0.11.1
+- Bump @uifabricshared/foundation-compose to v1.11.1
 
 ## 0.9.0