+++ packages/experimental/Shimmer/CHANGELOG.md
@@ -1,8 +1,42 @@
 # Change Log - @fluentui-react-native/experimental-shimmer
 
-This log was last generated on Wed, 17 Nov 2021 19:28:07 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Dec 2021 22:06:58 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.4
+
+Fri, 17 Dec 2021 22:06:58 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.4
+
+## 0.6.3
+
+Fri, 17 Dec 2021 19:53:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.3
+
+## 0.6.2
+
+Fri, 17 Dec 2021 01:26:42 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.2
+
+## 0.6.1
+
+Thu, 16 Dec 2021 19:10:27 GMT
+
+### Patches
+
+- Flip Shimmer in RTL (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/framework to v0.7.1
+- Bump @fluentui-react-native/tokens to v0.11.1
 
 ## 0.6.0