+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,42 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Wed, 17 Nov 2021 19:28:07 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Dec 2021 22:06:58 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.4
+
+Fri, 17 Dec 2021 22:06:58 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.10.3
+
+## 0.11.3
+
+Fri, 17 Dec 2021 19:53:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.10.2
+
+## 0.11.2
+
+Fri, 17 Dec 2021 01:26:42 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.10.1
+
+## 0.11.1
+
+Thu, 16 Dec 2021 19:10:27 GMT
+
+### Patches
+
+- Change focusRef type (ruaraki@microsoft.com)
+- Refactor useKeyCallback (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/icon to v0.10.0
 
 ## 0.11.0